# backend/documentation/specs/api-openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Early Harvest Viewer",
    "description": "Early Harvest Viewer API",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:3000/" }],
  "paths": {
...
    "/sites/{siteId}/multiple-data": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "description": "",
          "required": true,
          "schema": {
            "$id": "siteId",
            "description": "Site identifier",
            "type": "string",
            "minLength": 8,
            "maxLength": 36,
            "examples": ["db1de9ed-8368-45e5-ac9f-1506b44c318e"]
          }
        }
      ],
      "delete": {
        "tags": ["sites"],
        "summary": "",
        "description": "Multiple site data deleted",
        "operationId": "deleteSiteMultipleData",
        "responses": {
          "204": {
            "description": "Successfully deleted the site multiple data"
          },
          "400": {
            "description": "Bad request error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Bad request error",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "description": "The error code"
                    },
                    "error": {
                      "type": "string",
                      "description": "The error short message"
                    },
                    "message": {
                      "type": "string",
                      "description": "The error full message"
                    },
                    "messageCode": {
                      "type": "string",
                      "description": "The error message code"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Unauthorized error",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "description": "The error code"
                    },
                    "error": {
                      "type": "string",
                      "description": "The error short message"
                    },
                    "message": {
                      "type": "string",
                      "description": "The error full message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Permission required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Permission required",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "description": "The error code"
                    },
                    "error": {
                      "type": "string",
                      "description": "The error short message"
                    },
                    "message": {
                      "type": "string",
                      "description": "The error full message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Not found",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "description": "The error code"
                    },
                    "error": {
                      "type": "string",
                      "description": "The error short message"
                    },
                    "message": {
                      "type": "string",
                      "description": "The error full message"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Generic error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Generic error",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "description": "The error code"
                    },
                    "error": {
                      "type": "string",
                      "description": "The error short message"
                    },
                    "message": {
                      "type": "string",
                      "description": "The error full message"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Service Unavailable error",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "description": "The error code"
                    },
                    "error": {
                      "type": "string",
                      "description": "The error short message"
                    },
                    "message": {
                      "type": "string",
                      "description": "The error full message"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [],
        "deprecated": false,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {},
                "type": "object",
                "required": ["targetList"]
              }
            }
          }
        }
      }
    },
...
  },
...
  "security": [],
...
}